+++ src/vitePlugins.ts
@@ -1,78 +1,156 @@
-import * as fs from "fs";
+import * as fs from "node:fs/promises";
 import MagicString from "magic-string";
-import type { Plugin, PluginOption, UserConfig } from "vite";
-import { parse } from "@vue/compiler-sfc";
-import * as path from "path";
-import { rimrafSync } from "rimraf";
-import { toGleamPath, baseGeneratedGleamPath } from "./utils";
-import vitePluginGleam from "vite-gleam";
+import type { Plugin } from "vite";
+import {
+  gleamBuild,
+  toVleamGeneratedPath,
+  cleanGenerated,
+  readCompiledGleamFile,
+  gleamProjectName,
+  resolveGleamImport,
+  getGleamBlockFromCode,
+} from "./utils";
 
-export function vitePluginVueVleam(): PluginOption {
-  return [vitePluginVueGleamScript(), vitePluginGleam() as unknown as Plugin];
+export type VleamQuery = {};
+
+export function parseVleamRequest(id: string): {
+  filename: string;
+  query: VleamQuery;
+} {
+  const [filename] = id.split(`?`, 2);
+  return {
+    filename,
+    query: {},
+  };
 }
 
-export function vitePluginVueGleamScript(): Plugin {
-  const srcPath = path.join(process.cwd(), "./src");
+export async function vitePluginVueVleam(): Promise<Plugin> {
+  const projectRoot = process.cwd();
+  const projectName = await gleamProjectName(projectRoot);
+
+  const gleamScriptSfcs = new Set();
 
   return {
     name: "vite-plugin-vue-gleam-script",
     enforce: "pre",
-    buildStart(_options) {
-      rimrafSync(baseGeneratedGleamPath(srcPath));
+    async buildStart() {
+      await cleanGenerated(projectRoot);
+
+      await gleamBuild();
     },
-    async handleHotUpdate({ modules, server }) {
+    async resolveId(source, importer) {
+      if (!importer) {
+        return;
+      }
+
+      const { filename } = parseVleamRequest(importer);
+
+      const shouldResolve =
+        filename.endsWith(".gleam") ||
+        filename === "./gleam.mjs" ||
+        gleamScriptSfcs.has(filename);
+
+      if (!shouldResolve) {
+        return;
+      }
+
+      const resolvedGleamImport = await resolveGleamImport(
+        projectRoot,
+        filename,
+        source,
+        projectName,
+      );
+
+      if (resolvedGleamImport) {
+        return { id: resolvedGleamImport };
+      }
+    },
+    async handleHotUpdate({ modules, server, file }) {
       if (modules.some(({ info }) => info?.meta?.vueGleam?.isGleamScript)) {
         server.hot.send({
           type: "full-reload",
         });
         return [];
       }
+
+      const { filename } = parseVleamRequest(file);
+
+      if (filename.endsWith("gleam.mjs")) {
+        return [];
+      }
+
+      if (filename.endsWith(".gleam")) {
+        await gleamBuild();
+      }
+
+      return modules;
     },
-    transform(code, id) {
-      if (!id.endsWith(".vue")) {
-        return;
-      }
+    async transform(code, id) {
+      const { filename } = parseVleamRequest(id);
+      if (filename.endsWith(".vue")) {
+        const gleamTargetPath = await toVleamGeneratedPath(
+          projectRoot,
+          filename,
+        );
 
-      const { descriptor } = parse(code);
+        const gleamBlock = getGleamBlockFromCode(code);
 
-      const scriptBlock = descriptor.script;
+        if (!gleamBlock) {
+          return;
+        }
 
-      if (scriptBlock?.lang !== "gleam") {
-        return;
-      }
+        await fs.writeFile(gleamTargetPath, gleamBlock.content, {
+          encoding: "utf8",
+        });
 
-      const gleamTargetPath = toGleamPath(srcPath, id);
+        await gleamBuild();
 
-      fs.writeFileSync(gleamTargetPath, scriptBlock.content, {
-        encoding: "utf8",
-      });
+        const compiledContent = await readCompiledGleamFile(
+          projectRoot,
+          gleamTargetPath,
+          projectName,
+        );
 
-      const relativeGleamPath = path.relative(
-        path.dirname(id),
-        gleamTargetPath,
-      );
+        const magicString = new MagicString(code)
+          .replace(
+            gleamBlock.content,
+            compiledContent + "\nexport default default_export();",
+          )
+          .replace('lang="gleam"', 'lang="ts"');
 
-      const magicString = new MagicString(code)
-        .replace(
-          scriptBlock.content,
+        gleamScriptSfcs.add(filename);
 
-          `import { default_export } from "${relativeGleamPath}";\n` +
-            "export default default_export();\n",
-        )
-        .replace('lang="gleam"', 'lang="ts"');
-
-      return {
-        code: magicString.toString(),
-        map: magicString.generateMap({
-          source: id,
-          includeContent: true,
-        }),
-        meta: {
-          vueGleam: {
-            isGleamScript: true,
+        return {
+          code: magicString.toString(),
+          map: magicString.generateMap({
+            source: filename,
+            includeContent: true,
+          }),
+          meta: {
+            vueGleam: {
+              isGleamScript: true,
+            },
           },
-        },
-      };
+        };
+      } else if (filename.endsWith(".gleam")) {
+        const compiledContent = await readCompiledGleamFile(
+          projectRoot,
+          filename,
+          projectName,
+        );
+
+        const magicString = new MagicString(code)
+          .replace(code, compiledContent)
+          .replace('lang="gleam"', 'lang="ts"');
+
+        return {
+          code: magicString.toString(),
+          map: magicString.generateMap({
+            source: filename,
+            includeContent: true,
+          }),
+        };
+      }
     },
   } satisfies Plugin;
 }